data-collector/requirements.txt
# Data Collector Project Dependencies
# Core libraries for API specification processing
chromadb==0.4.22
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
lxml==4.9.3

# Data processing and utilities
pandas==2.1.4
numpy==1.24.3
markdown==3.5.2

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
